+++ source/AzAuth.PS/Cmdlets/GetAzToken.cs
@@ -79,7 +79,9 @@ public class GetAzToken : PSLoggerCmdletBase
     [ArgumentCompleter(typeof(ExistingAccounts))]
     public string Username { get; set; }
 
+    [Parameter(ParameterSetName = "NonInteractive")]
     [Parameter(ParameterSetName = "Interactive")]
+    [Parameter(ParameterSetName = "ManagedIdentity")]
     [Parameter(ParameterSetName = "DeviceCode")]
     [ValidateRange(1, int.MaxValue)]
     public int TimeoutSeconds { get; set; } = 120;
@@ -146,7 +148,17 @@ public class GetAzToken : PSLoggerCmdletBase
 
         if (ParameterSetName == "NonInteractive")
         {
+            // If user didn't specify a timeout, default to 1 second for managed identity
+            int managedIdentityTimeoutSeconds = 1;
+            int? noninteractiveTimeoutSeconds = null;
+            if (MyInvocation.BoundParameters.ContainsKey("TimeoutSeconds"))
+            {
+                managedIdentityTimeoutSeconds = TimeoutSeconds;
+                noninteractiveTimeoutSeconds = TimeoutSeconds;
+            }
+
             WriteVerbose(@"Looking for a token from the following sources:
+Managed Identity (https://learn.microsoft.com/en-us/dotnet/api/azure.identity.managedidentitycredential)
 Environment variables (https://learn.microsoft.com/en-us/dotnet/api/azure.identity.environmentcredential)
 Azure PowerShell (https://learn.microsoft.com/en-us/dotnet/api/azure.identity.azurepowershellcredential)
 Azure CLI (https://learn.microsoft.com/en-us/dotnet/api/azure.identity.azureclicredential)
@@ -154,7 +166,7 @@ Visual Studio Code (https://learn.microsoft.com/en-us/dotnet/api/azure.identity.
 Visual Studio (https://learn.microsoft.com/en-us/dotnet/api/azure.identity.visualstudiocredential)
 Shared token cache (https://learn.microsoft.com/en-us/dotnet/api/azure.identity.sharedtokencachecredential)
 ");
-            WriteObject(TokenManager.GetTokenNonInteractive(Resource, Scope, Claim, TenantId, stopProcessing.Token));
+            WriteObject(TokenManager.GetTokenNonInteractive(Resource, Scope, Claim, TenantId, noninteractiveTimeoutSeconds, managedIdentityTimeoutSeconds, stopProcessing.Token));
         }
         else if (ParameterSetName == "Cache")
         {
@@ -190,8 +202,13 @@ Shared token cache (https://learn.microsoft.com/en-us/dotnet/api/azure.identity.
         }
         else if (ManagedIdentity.IsPresent)
         {
+            // If user didn't specify a timeout, default to 1 second for managed identity
+            if (!MyInvocation.BoundParameters.ContainsKey("TimeoutSeconds"))
+            {
+                TimeoutSeconds = 1;
+            }
             WriteVerbose("Getting token using a managed identity (https://learn.microsoft.com/en-us/dotnet/api/azure.identity.managedidentitycredential).");
-            WriteObject(TokenManager.GetTokenManagedIdentity(Resource, Scope, Claim, ClientId, TenantId, stopProcessing.Token));
+            WriteObject(TokenManager.GetTokenManagedIdentity(Resource, Scope, Claim, ClientId, TenantId, TimeoutSeconds, stopProcessing.Token));
         }
         else if (WorkloadIdentity.IsPresent)
         {